chore: move image filter functions to src/image/image_filter.c
