+++ include/image/image_processing.h
@@ -8,10 +8,3 @@ void binarize_image(GdkPixbuf *pixbuf, int treshold);
 int convert_to_black_and_white(GdkPixbuf *pixbuf);
 int *create_histogram(GdkPixbuf *pixbuf);
 void print_histogram(GdkPixbuf *pixbuf);
-
-void erode_3x3(GdkPixbuf *pixbuf);
-void dilate_3x3(GdkPixbuf *pixbuf);
-
-int find_minimum_index(int start_index, guchar neighborhood[9]);
-void selection_sort(guchar neighborhood[9]);
-void median_filter_3x3(GdkPixbuf *pixbuf);
